* develop: (300 commits)
  0.28.19-alpha.4
  fix build hook pipes
  env var handlers
  crypto updates
  error handling
  0.28.19-alpha.2
  fix for absolute paths
  fonts fix
  iosphotoeditor
  update logging icon font excludes
  fonts, better hooks, plugins
  release 0.28.19-alpha.2
  removed superfluous command line parameter
  link works on whole package
  moves non-hoisted devdependencies up to top
  removing rnv direct dependency from deploy-docker
  adds back the link script
  0.28.19-alpha.1
  closes #398
  detox version up
  ...

# Conflicts:
#	package.json
#	packages/rnv/pluginTemplates/renative.plugins.json

+++ packages/rnv/src/systemTools/fileutils.js
@@ -6,12 +6,12 @@ import shelljs from 'shelljs';
 import merge from 'deepmerge';
 import chalk from 'chalk';
 import ncp from 'ncp';
+import { isSystemWin } from '../utils';
+
 import { logDebug, logError, logWarning, logInfo } from './logger';
 
-export const isRunningOnWindows = process.platform === 'win32';
-
 export const copyFileSync = (source, target) => {
-    logDebug('copyFileSync', source, target);
+    logDebug('copyFileSync', source);
     let targetFile = target;
     // if target is a directory a new file with the same name will be created
     if (source.indexOf('.DS_Store') !== -1) return;
@@ -21,7 +21,26 @@ export const copyFileSync = (source, target) => {
             targetFile = path.join(target, path.basename(source));
         }
     }
-    fs.writeFileSync(targetFile, fs.readFileSync(source));
+    if (fs.existsSync(targetFile)) {
+        const src = fs.readFileSync(source);
+        const dst = fs.readFileSync(targetFile);
+
+        if (Buffer.compare(src, dst) === 0) return;
+    }
+    logDebug('copyFileSync', source, targetFile, 'executed');
+    try {
+        fs.writeFileSync(targetFile, fs.readFileSync(source));
+    } catch (e) {
+        console.log('copyFileSync', e);
+    }
+};
+
+export const invalidatePodsChecksum = (c) => {
+    const appFolder = path.join(c.paths.project.builds.dir, `${c.runtime.appId}_${c.platform}`);
+    const podChecksumPath = path.join(appFolder, 'Podfile.checksum');
+    if (fs.existsSync(podChecksumPath)) {
+        fs.unlinkSync(podChecksumPath);
+    }
 };
 
 export const copyFolderRecursiveSync = (source, target, convertSvg = true, skipPaths) => {
@@ -184,12 +203,26 @@ export const removeDirSync = (dir, rmSelf) => {
     }
 };
 
-export const writeObjectSync = (filePath, obj, spaces, addNewLine = true) => {
-    if (addNewLine) {
-        fs.writeFileSync(filePath, `${JSON.stringify(obj, null, spaces || 4)}\n`);
+export const writeFileSync = (filePath, obj, spaces, addNewLine = true) => {
+    logDebug('writeFileSync', filePath);
+    if (filePath.includes('?') || filePath.includes('undefined')) return;
+    let output;
+    if (typeof obj === 'string') {
+        output = obj;
     } else {
-        fs.writeFileSync(filePath, JSON.stringify(obj, null, spaces || 4));
+        output = `${JSON.stringify(obj, null, spaces || 4)}${addNewLine ? '\n' : ''}`;
     }
+    if (fs.existsSync(filePath)) {
+        if (fs.readFileSync(filePath).toString() === output) return;
+    }
+    logDebug('writeFileSync', filePath, 'executed');
+    fs.writeFileSync(filePath, output);
+};
+
+export const writeObjectSync = (filePath, obj, spaces, addNewLine = true) => {
+    logDebug('writeObjectSync', filePath);
+    logWarning('writeObjectSync is DEPRECATED. use writeFileSync instead');
+    return writeFileSync(filePath, obj, spaces, addNewLine);
 };
 
 export const readObjectSync = (filePath, sanitize = false, c) => {
@@ -215,7 +248,7 @@ export const readObjectSync = (filePath, sanitize = false, c) => {
             }
         }
     } catch (e) {
-        logWarning(`readObjectSync: Parsing of ${chalk.white(filePath)} failed with ${e}`);
+        logError(`readObjectSync: Parsing of ${chalk.white(filePath)} failed with ${e}`);
         return null;
     }
     return obj;
@@ -229,7 +262,7 @@ export const updateObjectSync = (filePath, updateObj) => {
     } else {
         output = updateObj;
     }
-    writeObjectSync(filePath, output);
+    writeFileSync(filePath, output);
     return output;
 };
 
@@ -310,7 +343,7 @@ export const sanitizeDynamicProps = (obj, props) => {
         obj.forEach((v) => {
             if (typeof val === 'string') {
                 Object.keys(props).forEach((pk) => {
-                    val = val.replace(`@${pk}@`, props[pk]);
+                    val = val.replace(`@${pk}@`, props[pk]).replace(`{{props.${pk}}}`, props[pk]);
                     obj[key] = val;
                 });
             } else {
@@ -323,7 +356,7 @@ export const sanitizeDynamicProps = (obj, props) => {
         if (val) {
             if (typeof val === 'string') {
                 Object.keys(props).forEach((pk) => {
-                    val = val.replace(`@${pk}@`, props[pk]);
+                    val = val.replace(`@${pk}@`, props[pk]).replace(`{{props.${pk}}}`, props[pk]);
                     obj[key] = val;
                 });
             } else {
@@ -333,6 +366,7 @@ export const sanitizeDynamicProps = (obj, props) => {
     });
     return obj;
 };
+
 
 export const mergeObjects = (c, obj1, obj2, dynamicRefs = true, replaceArrays = false) => {
     if (!obj2) return obj1;
@@ -362,7 +396,7 @@ export const updateConfigFile = async (update, globalConfigPath) => {
 };
 
 export const replaceHomeFolder = (p) => {
-    if (isRunningOnWindows) return p.replace('~', process.env.USERPROFILE);
+    if (isSystemWin) return p.replace('~', process.env.USERPROFILE);
     return p.replace('~', process.env.HOME);
 };
 
@@ -397,7 +431,7 @@ export default {
     copyFolderContentsRecursive,
     copyFolderContentsRecursiveSync,
     cleanFolder,
-    writeObjectSync,
+    writeFileSync,
     readObjectSync,
     updateObjectSync,
     arrayMerge,
